freebasic-bootstrap: Bump version to ppa2 (oops)
The ppa1 upload had an invalid distro, and Launchpad refused it. All I had
to do was to fix the distro and reupload. But dput refused to reupload
because the *.upload file still existed. So I thought it was necessary
to bump the version, but it wasn't. But now I'll keep it, because it's
uploaded, and I don't want to delete it again.