Make unified freebasic-bootstrap package, use upstream FB tarball
Now there is just one freebasic-bootstrap package, instead of one per
target.

Also, the FB sources are now simply extracted from the upstream tarball,
instead of being retrieved from Git. (The problem is that we can't recreate
the upstream tarball during every package build, because the timestamps
change and cause the tarball to be different, despite having the same name)